package strcture changed to isolate the movie domain from the user domain
