+++ movie/src/main/java/com/entertainment/MovieApplication.java
@@ -1,0 +1,13 @@
+package com.entertainment;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class MovieApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(MovieApplication.class, args);
+	}
+
+}
