+++ movie/src/main/java/com/entertainment/domain/movie/exception/MovieNotFoundException.java
@@ -1,0 +1,8 @@
+package com.entertainment.domain.movie.exception;
+
+public class MovieNotFoundException extends RuntimeException{
+
+    public MovieNotFoundException(String uuid) {
+        super("Could not find movie " + uuid + ".");
+    }
+}
