+++ movie/src/main/java/com/entertainment/domain/movie/exception/RateCreateException.java
@@ -1,0 +1,8 @@
+package com.entertainment.domain.movie.exception;
+
+public class RateCreateException extends RuntimeException {
+
+    public RateCreateException(String message) {
+        super("An error occurred while trying save the rating: " + message);
+    }
+}
