+++ movie/src/main/java/com/entertainment/domain/movie/outgoing/RatingStorageService.java
@@ -1,0 +1,8 @@
+package com.entertainment.domain.movie.outgoing;
+
+import com.entertainment.domain.movie.core.Rating;
+
+public interface RatingStorageService {
+
+    void saveRating(Rating rating);
+}
