+++ movie/src/main/java/com/entertainment/domain/user/service/UserDetailService.java
@@ -1,0 +1,4 @@
+package com.entertainment.domain.user.service;
+
+public class UserDetailService {
+}
