--- movie/src/main/java/com/entertainment/movie/MovieApplication.java
@@ -1,13 +1,0 @@
-package com.entertainment.movie;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class MovieApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(MovieApplication.class, args);
-	}
-
-}
