--- movie/src/main/java/com/entertainment/movie/domain/exception/MovieCreateException.java
@@ -1,8 +1,0 @@
-package com.entertainment.movie.domain.exception;
-
-public class MovieCreateException extends RuntimeException{
-
-    public MovieCreateException(String message) {
-        super("An error occurred while trying create the movie: " + message);
-    }
-}
