--- movie/src/main/java/com/entertainment/movie/domain/exception/MovieNotFoundException.java
@@ -1,8 +1,0 @@
-package com.entertainment.movie.domain.exception;
-
-public class MovieNotFoundException extends RuntimeException{
-
-    public MovieNotFoundException(String uuid) {
-        super("Could not find movie " + uuid + ".");
-    }
-}
