--- movie/src/main/java/com/entertainment/movie/domain/exception/RateCreateException.java
@@ -1,8 +1,0 @@
-package com.entertainment.movie.domain.exception;
-
-public class RateCreateException extends RuntimeException {
-
-    public RateCreateException(String message) {
-        super("An error occurred while trying save the rating: " + message);
-    }
-}
