--- movie/src/main/java/com/entertainment/movie/domain/outgoing/RatingStorageService.java
@@ -1,8 +1,0 @@
-package com.entertainment.movie.domain.outgoing;
-
-import com.entertainment.movie.domain.service.core.Rating;
-
-public interface RatingStorageService {
-
-    void saveRating(Rating rating);
-}
