--- movie/src/main/java/com/entertainment/movie/domain/service/core/Rating.java
@@ -1,21 +1,0 @@
-package com.entertainment.movie.domain.service.core;
-
-import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
-
-import java.util.Date;
-import java.util.UUID;
-
-@Builder
-@Getter
-@Setter
-public class Rating {
-    private UUID id;
-    private UUID movieId;
-    private UUID userId;
-    private int ratingValue;
-    private String comment;
-    private Date createdAt;
-    private Date updatedAt;
-}
