--- movie/src/main/java/com/entertainment/movie/port/outgoing/db/repository/MovieRepository.java
@@ -1,9 +1,0 @@
-package com.entertainment.movie.port.outgoing.db.repository;
-
-import com.entertainment.movie.port.outgoing.db.entity.MovieEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.UUID;
-
-public interface MovieRepository extends JpaRepository<MovieEntity, UUID> {
-}
