+++ movie/src/main/java/com/entertainment/port/outgoing/db/entity/UserEntity.java
@@ -1,0 +1,32 @@
+package com.entertainment.port.outgoing.db.entity;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.util.Date;
+import java.util.UUID;
+
+@Entity(name = "user_detail")
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+@Getter
+public class UserEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private UUID id;
+    private String userName;
+    private String firstName;
+    private String lastName;
+    private String email;
+    private String address;
+    private Date createdAt;
+    private Date updatedAt;
+}
