+++ movie/src/main/resources/api/movie-api-spec.yaml
@@ -144,9 +144,75 @@ paths:
             application/json:
               schema:
                 $ref: '#/components/schemas/CommonResponse'
+  /user:
+    post:
+      operationId: createUser
+      tags:
+        - user
+      requestBody:
+        required: true
+        content:
+          application/json:
+            schema:
+              $ref: '#/components/schemas/UserDto'
+
+      responses:
+        '200':
+          description: Successful response.
+          content:
+            application/json:
+              schema:
+                $ref: '#/components/schemas/CommonResponse'
+        '400':
+          description: Bad request.
+          content:
+            application/json:
+              schema:
+                $ref: '#/components/schemas/CommonResponse'
+        '500':
+          description: Internal server error response.
+          content:
+            application/json:
+              schema:
+                $ref: '#/components/schemas/CommonResponse'
 
 components:
   schemas:
+    UserDto:
+      type: object
+      properties:
+        userName:
+          type: string
+          description: "user ID of the user"
+          minLength: 2
+          maxLength: 30
+        firstName:
+          type: string
+          description: "first name of the user"
+          minLength: 2
+          maxLength: 100
+        lastName:
+          type: string
+          description: "last name of the user"
+          minLength: 2
+          maxLength: 100
+        email:
+          type: string
+          format: email
+          description: "email address"
+          minLength: 3
+          maxLength: 150
+        address:
+          type: string
+          description: "residence address"
+          minLength: 10
+          maxLength: 255
+      required:
+        - userId
+        - firstName
+        - lastName
+        - email
+        - address
     UserRatedMovieList:
       type: array
       items:
