+++ movie/src/main/resources/db/migration/V1_6__initialize_user_table.sql
@@ -1,0 +1,12 @@
+CREATE TABLE IF NOT EXISTS user_detail (
+    id UUID NOT NULL,
+    user_name VARCHAR(30) NOT NULL,
+    first_name VARCHAR(100) NOT NULL,
+    last_name VARCHAR(100) NOT NULL,
+    email VARCHAR(150) NOT NULL,
+    address VARCHAR(255) NOT NULL,
+    PRIMARY KEY (id)
+);
+
+
+
