+++ movie/src/test/java/com/entertainment/movie/port/incoming/MovieApiIntegrationTest.java
@@ -4,8 +4,8 @@ import com.entertainment.config.TestExtension;
 import com.entertainment.movie.dto.BasicMovieDto;
 import com.entertainment.movie.dto.CommonResponse;
 import com.entertainment.movie.dto.ExtendedMovieDto;
-import com.entertainment.movie.port.outgoing.db.entity.MovieEntity;
-import com.entertainment.movie.port.outgoing.db.repository.MovieRepository;
+import com.entertainment.port.outgoing.db.entity.MovieEntity;
+import com.entertainment.port.outgoing.db.repository.MovieRepository;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
