+++ movie/src/test/java/com/entertainment/movie/port/incoming/RateApiIntegrationTest.java
@@ -3,10 +3,10 @@ package com.entertainment.movie.port.incoming;
 import com.entertainment.config.TestExtension;
 import com.entertainment.movie.dto.CommonResponse;
 import com.entertainment.movie.dto.RateDto;
-import com.entertainment.movie.port.outgoing.db.entity.MovieEntity;
-import com.entertainment.movie.port.outgoing.db.entity.RatingEntity;
-import com.entertainment.movie.port.outgoing.db.repository.MovieRepository;
-import com.entertainment.movie.port.outgoing.db.repository.RatingRepository;
+import com.entertainment.port.outgoing.db.entity.MovieEntity;
+import com.entertainment.port.outgoing.db.entity.RatingEntity;
+import com.entertainment.port.outgoing.db.repository.MovieRepository;
+import com.entertainment.port.outgoing.db.repository.RatingRepository;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
